requirements.txt
pandas
sqlalchemy
python-dotenv
pymysql
psycopg2-binary